+++ pkg/chain/chain.go
@@ -218,7 +218,7 @@ func (c *Chain) GetAccountLocationsInTxs(ctx context.Context, txs []transaction.
 	}
 
 	for i, requestAccount := range requestAccounts {
-		accountLocations[requestAccount] = states[i].ShardLocation
+		accountLocations[requestAccount] = int64(states[i].ShardLocation)
 	}
 
 	return accountLocations, nil
@@ -493,7 +493,7 @@ func (c *Chain) updateStateMapByRelayTx(accountStates map[account.Address]*accou
 	case transaction.Relay1Tx:
 		// For a relay1 transaction, debit the sender's balance.
 		senderState := accountStates[tx.Sender]
-		if senderState == nil || senderState.ShardLocation != c.shardID {
+		if senderState == nil || senderState.ShardLocation != uint64(c.shardID) {
 			// Sender is not in this shard, skip.
 			return
 		}
@@ -507,7 +507,7 @@ func (c *Chain) updateStateMapByRelayTx(accountStates map[account.Address]*accou
 	case transaction.Relay2Tx:
 		// For a relay2 transaction credit the recipient's balance.
 		recipientState := accountStates[tx.Recipient]
-		if recipientState == nil || recipientState.ShardLocation != c.shardID {
+		if recipientState == nil || recipientState.ShardLocation != uint64(c.shardID) {
 			return
 		}
 
@@ -525,7 +525,7 @@ func (c *Chain) modifyStateMapByBrokerTx(accountStates map[account.Address]*acco
 		senderState := accountStates[tx.Sender]
 
 		brokerState := accountStates[tx.Broker]
-		if senderState == nil || senderState.ShardLocation != c.shardID {
+		if senderState == nil || senderState.ShardLocation != uint64(c.shardID) {
 			slog.Error("handle broker1 tx error", "err", "the sender is not in this shard", "sender", tx.Sender, "shard", c.shardID)
 			return
 		}
@@ -542,7 +542,7 @@ func (c *Chain) modifyStateMapByBrokerTx(accountStates map[account.Address]*acco
 		recipientState := accountStates[tx.Recipient]
 
 		brokerState := accountStates[tx.Broker]
-		if recipientState == nil || recipientState.ShardLocation != c.shardID {
+		if recipientState == nil || recipientState.ShardLocation != uint64(c.shardID) {
 			slog.Error("handle broker2 tx error", "err", "the recipient is not in this shard", "recipient", tx.Recipient, "shard", c.shardID)
 			return
 		}
@@ -564,7 +564,7 @@ func (c *Chain) modifyStateMapByNormalTx(accountStates map[account.Address]*acco
 	recipientState := accountStates[tx.Recipient]
 
 	// Modify senderState
-	if senderState != nil && senderState.ShardLocation != c.shardID {
+	if senderState != nil && senderState.ShardLocation != uint64(c.shardID) {
 		if err := senderState.Debit(tx.Value); errors.Is(err, account.ErrNotEnoughBalance) {
 			slog.Warn("the balance of sender is not enough", "sender", tx.Sender, "value", tx.Value)
 		} else if err != nil {
@@ -573,7 +573,7 @@ func (c *Chain) modifyStateMapByNormalTx(accountStates map[account.Address]*acco
 	}
 
 	// Modify recipientState
-	if recipientState != nil && recipientState.ShardLocation != c.shardID {
+	if recipientState != nil && recipientState.ShardLocation != uint64(c.shardID) {
 		recipientState.Credit(tx.Value)
 	}
 }
